restructured file structure
