+++ data_profiler/helpers/constants/app_constants.py
@@ -1,0 +1,15 @@
+'''
+Jack Miller
+Apex Companies
+Oct 2024
+
+General Python constants for DataProfiler app
+'''
+
+
+''' General '''
+
+# Resources
+RESOURCES_DIR = './resources'
+RESOURCES_DIR_DEV = './resources'
+
