+++ data_profiler/helpers/constants/data_file_constants.py
@@ -1,31 +1,13 @@
+'''
+Jack Miller
+Apex Companies
+Oct 2024
 
-from enum import Enum
-
-
-###############
-### General ###
-###############
-
-# Resources
-RESOURCES_DIR = './resources'
-RESOURCES_DIR_DEV = './resources'
+Python constants relating to input/upload files for DataProfiler
+'''
 
 
-
-
-
-
-######################
-### Uploaded Files ###
-######################
-
-class UploadFileTypes(str, Enum):
-    ITEM_MASTER = 'ItemMaster'
-    INBOUND_HEADER = 'InboundHeader'
-    INBOUND_DETAILS = 'InboundDetails'
-    INVENTORY = 'Inventory'
-    ORDER_HEADER = 'OrderHeader'
-    ORDER_DETAILS = 'OrderDetails'
+''' Required Columns '''
 
 ITEM_MASTER_UPLOAD_REQUIRED_COLS = ['SKU','SKUDescription','SKUClass','UnitOfMeasure','EachLength','EachWidth','EachHeight','EachWeight','InnerQuantity','InnerLength','InnerWidth','InnerHeight','InnerWeight','CartonQuantity','CartonLength','CartonWidth','CartonHeight','CartonWeight','CartonsPerPallet','PalletTie','PalletHigh','MaxPalletStack','PalletLength','PalletWidth','PalletHeight','PalletWeight','Conveyable','Subwarehouse','AllowToPickPallet','AllowToPickCarton','AllowToPickInnerPacks','AllowToPickUnits']
 INBOUND_HEADER_UPLOAD_REQUIRED_COLS = ['ReceiptNumber', 'ArrivalDate', 'ArrivalTime', 'ExpectedDate', 'ExpectedTime', 'Carrier', 'Mode']
@@ -42,6 +24,9 @@ UPLOADS_REQUIRED_COLUMNS_MAPPER = {
     'OrderHeader': ORDER_HEADER_UPLOAD_REQUIRED_COLS,
     'OrderDetails': ORDER_DETAILS_UPLOAD_REQUIRED_COLS
 }
+
+
+''' Required column types '''
 
 ITEM_MASTER_UPLOAD_REQUIRED_DTYPES = {
     'SKU':'object',
@@ -103,7 +88,6 @@ INVENTORY_UPLOAD_REQUIRED_DTYPES = {
 }
 
 ORDER_HEADER_UPLOAD_REQUIRED_DTYPES = {
-    # 'ProjectNumber': 'object',
     'OrderNumber':'object',
     'ReceivedDate':'date',
     'PickDate':'date',
@@ -112,7 +96,6 @@ ORDER_HEADER_UPLOAD_REQUIRED_DTYPES = {
 }
 
 ORDER_DETAILS_UPLOAD_REQUIRED_DTYPES = {
-    # 'ProjectNumber': 'object',
     'OrderNumber':'object',
     'SKU':'object',
     'Quantity':'int64',
@@ -139,3 +122,28 @@ DTYPES_DEFAULT_VALUES = {
     'int64': 0,
     'float64': 0.0
 }
+
+
+''' File Errors '''
+
+DIRECTORY_ERROR_DOES_NOT_EXIST = 'The given data directory does not exist.'
+
+FILE_ERROR_MISSING_ITEM_MASTER = 'Item Master is missing. Cannot continue.'
+FILE_ERROR_MISSING_INBOUND_HEADER = '"Process Inbound Data" set to true but Inbound Header is not found.'
+FILE_ERROR_MISSING_INBOUND_DETAILS = '"Process Inbound Data" set to true but Inbound Details is not found.'
+FILE_ERROR_MISSING_INVENTORY = '"Process Inventory Data" set to true but Inventory file is not found.'
+FILE_ERROR_MISSING_OUTBOUND_HEADER = '"Process Outbound Data" set to true but Outbound Header is not found.'
+FILE_ERROR_MISSING_OUTBOUND_DETAILS = '"Process Outbound Data" set to true but Outbound Details is not found.'
+
+FILE_ERROR_ITEM_MASTER_MISSING_COLUMNS = 'Item Master is missing columns.'
+FILE_ERROR_INBOUND_HEADER_MISSING_COLUMNS = 'Inbound Header is missing columns.'
+FILE_ERROR_INBOUND_DETAILS_MISSING_COLUMNS = 'Inbound Details is missing columns.'
+FILE_ERROR_INVENTORY_MISSING_COLUMNS = 'Inventory is missing columns.'
+FILE_ERROR_ORDER_HEADER_MISSING_COLUMNS = 'Order Header is missing columns.'
+FILE_ERROR_ORDER_DETAILS_MISSING_COLUMNS = 'Order Details is missing columns.'
+
+# Unused
+FILE_ERROR_INBOUND_NO_HEADER = 'Inbound Details was provided without Inbound Header.'
+FILE_ERROR_INBOUND_NO_DETAILS = 'Inbound Header was provided without Inbound Details.'
+FILE_ERROR_OUTBOUND_NO_HEADER = 'Outbound Details was provided without Outbound Header.'
+FILE_ERROR_OUTBOUND_NO_DETAILS = 'Outbound Header was provided without Outbound Details.'
